+++ include/ecc/ecc_key.h
@@ -8,7 +8,7 @@
 struct ecc_key {
     struct ecc_point p;
     struct integer_st d;
-    int priv;
+    int8_t priv;
 };
 
 void ecc_key_data_init(struct ecc_key *);
@@ -16,7 +16,7 @@ void ecc_key_data_free(struct ecc_key *);
 
 void ecc_key_clear(struct ecc_key *);
 
-int ecc_key_set_str(struct ecc_key *, const struct string_st *, const struct ecc_curve *);
+int8_t ecc_key_set_str(struct ecc_key *, const struct string_st *, const struct ecc_curve *);
 void ecc_key_get_str(const struct ecc_key *, struct string_st *);
 void ecc_key_get_address(const struct ecc_key *, struct string_st *);
