+++ include/ecc/ecc_point.h
@@ -16,7 +16,7 @@ void ecc_point_data_free(struct ecc_point *);
 void ecc_point_set(struct ecc_point *, const struct ecc_point *);
 void ecc_point_clear(struct ecc_point *);
 
-int ecc_point_set_str(struct ecc_point *, const struct string_st *tlv, const struct ecc_curve *curve);
+int8_t ecc_point_set_str(struct ecc_point *, const struct string_st *tlv, const struct ecc_curve *curve);
 void ecc_point_get_str(const struct ecc_point *P, struct string_st *);
 
 void ecc_point_double(struct ecc_point *, const struct ecc_point *P, const struct ecc_curve *curve);
